+++ dev-infra/commit-message/restore-commit-message/restore-commit-message.ts
@@ -11,7 +11,7 @@ import {writeFileSync} from 'fs';
 import {debug, log} from '../../utils/console';
 
 import {loadCommitMessageDraft} from '../commit-message-draft';
-import {CommitMsgSource} from '../config';
+import {CommitMsgSource} from '../commit-message-source';
 
 /**
  * Restore the commit message draft to the git to be used as the default commit message.
